refactor ready
